+++ lib/datasafe/lib/model/data-safe-private-endpoint-summary.ts
@@ -54,6 +54,26 @@ export interface DataSafePrivateEndpointSummary {
    * The current state of the private endpoint.
    */
   "lifecycleState"?: model.LifecycleState;
+  /**
+    * Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace. For more information, see [Resource Tags](https://docs.cloud.oracle.com/iaas/Content/General/Concepts/resourcetags.htm)
+* <p>
+Example: `{\"Department\": \"Finance\"}`
+* 
+    */
+  "freeformTags"?: { [key: string]: string };
+  /**
+    * Defined tags for this resource. Each key is predefined and scoped to a namespace. For more information, see [Resource Tags](https://docs.cloud.oracle.com/iaas/Content/General/Concepts/resourcetags.htm)
+* <p>
+Example: `{\"Operations\": {\"CostCenter\": \"42\"}}`
+* 
+    */
+  "definedTags"?: { [key: string]: { [key: string]: any } };
+  /**
+   * System tags for this resource. Each key is predefined and scoped to a namespace. For more information, see Resource Tags.
+   * Example: `{\"orcl-cloud\": {\"free-tier-retained\": \"true\"}}`
+   *
+   */
+  "systemTags"?: { [key: string]: { [key: string]: any } };
 }
 
 export namespace DataSafePrivateEndpointSummary {