+++ lib/managementdashboard/lib/model/update-management-dashboard-details.ts
@@ -89,6 +89,10 @@ export interface UpdateManagementDashboardDetails {
    */
   "isFavorite"?: boolean;
   /**
+   * Defines parameters for the dashboard.
+   */
+  "parametersConfig"?: Array<any>;
+  /**
    * Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
    * Example: `{\"bar-key\": \"value\"}`
    *